requirements.txt
#python 3.11.11
accelerate==1.6.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
cairocffi==1.7.1
CairoSVG==2.8.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cramjam==2.10.0
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
diffusers==0.33.1
fastparquet==2024.11.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
ftfy==6.3.1
google==3.0.0
grpcio==1.71.0
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.0
kagglehub==0.3.12
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
pandas==2.2.3
peft==0.15.2
pillow==11.2.1
polars==1.29.0
protobuf==6.30.2
pyarrow==20.0.0
pycparser==2.22
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
sentencepiece==0.2.0
soupsieve==2.7
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1